+++ solvers/CardiacEPSolver/CellModels/Fox02.cpp
@@ -74,7 +74,7 @@ void Fox02::v_Update(const Array<OneD, const Array<OneD, NekDouble>> &inarray,
 {
     boost::ignore_unused(time);
 
-    for (unsigned int i = 0; i < m_nq; ++i)
+    for (size_t i = 0; i < m_nq; ++i)
     {
 
         // Inputs:
